% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{context_create}
\alias{context_create}
\title{Create inference context}
\usage{
context_create(model, n_ctx = 2048L, n_threads = 4L, n_seq_max = 1L)
}
\arguments{
\item{model}{A model object returned by model_load()}

\item{n_ctx}{Context size (default: 2048)}

\item{n_threads}{Number of threads (default: 4)}

\item{n_seq_max}{Maximum number of sequences (default: 1)}
}
\value{
A context object (external pointer)
}
\description{
Create inference context
}
