% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{detokenize}
\alias{detokenize}
\title{Detokenize tokens}
\usage{
detokenize(model, tokens)
}
\arguments{
\item{model}{A model object}

\item{tokens}{Integer vector of token IDs}
}
\value{
Detokenized text string
}
\description{
Detokenize tokens
}
