% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.check_model_memory_requirements}
\alias{.check_model_memory_requirements}
\title{Check memory requirements for model loading}
\usage{
.check_model_memory_requirements(model_path)
}
\arguments{
\item{model_path}{Path to the model file}
}
\description{
Check memory requirements for model loading
}
