% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.download_model_to_cache}
\alias{.download_model_to_cache}
\title{Download a model to cache}
\usage{
.download_model_to_cache(model_url, cache_path, show_progress = TRUE)
}
\arguments{
\item{model_url}{The model URL}

\item{cache_path}{The local cache path}

\item{show_progress}{Whether to show download progress}
}
\description{
Download a model to cache
}
