% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_llama.R
\name{.ensure_model_loaded}
\alias{.ensure_model_loaded}
\title{Ensure model and context are loaded}
\usage{
.ensure_model_loaded(model_path, n_gpu_layers, n_ctx, n_threads)
}
\arguments{
\item{model_path}{Model path or URL}

\item{n_gpu_layers}{Number of GPU layers}

\item{n_ctx}{Context size}

\item{n_threads}{Number of threads}
}
\description{
Ensure model and context are loaded
}
