% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_llama.R
\name{.is_backend_loaded}
\alias{.is_backend_loaded}
\title{Check if backend is loaded}
\usage{
.is_backend_loaded()
}
\value{
TRUE if backend is loaded, FALSE otherwise
}
\description{
Check if backend is loaded
}
