% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.is_url}
\alias{.is_url}
\title{Check if a string represents a URL}
\usage{
.is_url(path)
}
\arguments{
\item{path}{The path/URL to check}
}
\value{
TRUE if it's a URL, FALSE otherwise
}
\description{
Check if a string represents a URL
}
