% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_model_cache_dir}
\alias{get_model_cache_dir}
\title{Get the model cache directory}
\usage{
get_model_cache_dir()
}
\value{
Path to the directory where models are cached
}
\description{
Get the model cache directory
}
