% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize text}
\usage{
tokenize(model, text, add_special = TRUE)
}
\arguments{
\item{model}{A model object}

\item{text}{Text to tokenize}

\item{add_special}{Whether to add special tokens (default: TRUE)}
}
\value{
Integer vector of token IDs
}
\description{
Tokenize text
}
