% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tokenize_test}
\alias{tokenize_test}
\title{Test tokenize function (debugging)}
\usage{
tokenize_test(model)
}
\arguments{
\item{model}{A model object}
}
\value{
Integer vector of tokens for "H"
}
\description{
Test tokenize function (debugging)
}
